requirements.txt
flask
google-cloud-vision
google-generativeai
pytesseract
pillow
pytest
coverage